Allow worker's group and development passwords.
This is an improvement for 'create_worker()' function, which allows
to use a single group password for a set of the workers instead of
a worker specific password.

Also it provides a default development password for the workers with
no password specified in case the buildbot instance is configured as
a development instance.